Answer what an action did to the displayed book outside the book
The level diff, the order changes and the trade prints move out of OrderBook
into DisplayedBookReport. None of it is about matching: what it needs is two
level windows, an event list and an instant, and what it produces is three kinds
of event. OrderBook drops from 1481 lines to 1253.

The seam is narrow on purpose. The report holds no reference to a book, a
matcher or a ladder between calls - the book supplies the ladders on each of the
two calls, because only the book has them, and that is the whole of the
coupling. It keeps the four scratch windows and the depths, since those are its
own working state and nothing else read them.

CollectLevelChanges comes out static and internal rather than private, which is
the point of the move: two windows and a depth in, a set of changes out. The
case that makes price-keyed reporting necessary - a level pushed out of a
five-deep window that appears nowhere in the ten-deep report - has never had a
direct test, because reaching it meant driving order flow through a whole book
and reading the tail of an event list. It can be written now, and is the next
commit.

Depth validation moves with the depths. Same messages, same parameter names,
same exception types, which is what PublishedDepthTests asserts.

No behaviour changes: same events, same order, same contents. The book still
owns which depths it reports at and still exposes AlsoReport, both of which now
forward. Moving that ownership to the group - and closing the case where a
pre-built book and its channel disagree, which InstrumentGroup documents rather
than prevents - is a separate change.
